requirements.txt
asgiref==3.2.7
attrs==19.3.0
certifi==2020.4.5.2
chardet==3.0.4
coverage==5.1
Django==3.0.7
django-cors-headers==3.3.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
httpie==2.1.0
idna==2.9
importlib-metadata==1.6.1
more-itertools==8.3.0
packaging==20.4
pluggy==0.13.1
py==1.8.1
Pygments==2.6.1
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.9.0
pytest-django==3.9.0
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
wcwidth==0.2.4
zipp==3.1.0
